+++ packages/omgidl-parser/src/IdlNodes/IdlNode.ts
@@ -1,12 +1,10 @@
-import { ConstantValue } from "@foxglove/message-definition";
+import { ConstantIdlNode } from "./ConstantIdlNode";
+import { BaseAstNode } from "../astTypes";
 
-import { BaseAstNode, ConstantAstNode, EnumAstNode } from "../astTypes";
-import { SIMPLE_TYPES, normalizeType } from "../primitiveTypes";
-import { IdlMessageDefinition, IdlMessageDefinitionField } from "../types";
-
-/** NOTE: All of the classes in this file are included such that we don't have circular import issues */
-
-/** Class used to resolve ASTNodes to IDLMessageDefinitions */
+/** Class used to resolve ASTNodes to IDLMessageDefinitions
+ * There is a subclass of this class for each `declarator` type on the AstNode.
+ * This class is meant to provide functions and variables that exist across all of the nodes to help them resolve to complete message definitions.
+ */
 export class IdlNode<T extends BaseAstNode = BaseAstNode> {
   /** Map of all IdlNodes in a schema definition */
   private map: Map<string, IdlNode>;
@@ -65,92 +63,6 @@ export class IdlNode<T extends BaseAstNode = BaseAstNode> {
   }
 }
 
-/** Wraps constant node so that its type and value can be resolved and written to a message definition */
-export class ConstantIdlNode extends IdlNode<ConstantAstNode> {
-  /** If the type needs resolution (not simple primitive) this will be set to true. Should only ever mean that it's referencing an enum */
-  private typeNeedsResolution = false;
-  constructor(scopePath: string[], astNode: ConstantAstNode, idlMap: Map<string, IdlNode>) {
-    super(scopePath, astNode, idlMap);
-    if (!SIMPLE_TYPES.has(astNode.type)) {
-      this.typeNeedsResolution = true;
-    }
-  }
-
-  get type(): string {
-    if (this.typeNeedsResolution) {
-      return this.getReferencedEnumNode().type;
-    }
-    return this.astNode.type;
-  }
-
-  /** Holds reference so that it doesn't need to be searched for again */
-  private referencedEnumNode?: EnumIdlNode = undefined;
-  /** Gets enum node referenced by type. Fails otherwise. */
-  private getReferencedEnumNode(): EnumIdlNode {
-    if (this.referencedEnumNode == undefined) {
-      const maybeEnumNode = this.getNode(this.scopePath, this.astNode.type);
-      if (!(maybeEnumNode instanceof EnumIdlNode)) {
-        throw new Error(`Expected ${this.astNode.type} to be an enum in ${this.scopedIdentifier}`);
-      }
-      this.referencedEnumNode = maybeEnumNode;
-    }
-    return this.referencedEnumNode;
-  }
-
-  get isConstant(): true {
-    return true;
-  }
-
-  /** Return Literal value on astNode or if the constant references another constant, then it gets the value that constant uses */
-  get value(): ConstantValue {
-    if (typeof this.astNode.value === "object") {
-      return this.getConstantNode(this.astNode.value.name).value;
-    }
-    return this.astNode.value;
-  }
-
-  /** Writes resolved IdlMessageDefinition */
-  toIDLMessageDefinitionField(): IdlMessageDefinitionField {
-    return {
-      name: this.name,
-      type: normalizeType(this.type),
-      value: this.value,
-      isConstant: true,
-      isComplex: false,
-      ...(this.astNode.valueText != undefined ? { valueText: this.astNode.valueText } : undefined),
-    };
-  }
-}
-
-/** Class used to resolve an Enum ASTNode to an IdlMessageDefinition */
-export class EnumIdlNode extends IdlNode<EnumAstNode> {
-  constructor(scopePath: string[], astNode: EnumAstNode, idlMap: Map<string, IdlNode>) {
-    super(scopePath, astNode, idlMap);
-  }
-
-  get type(): string {
-    return "uint32";
-  }
-
-  private enumeratorNodes(): ConstantIdlNode[] {
-    return this.astNode.enumerators.map((enumerator) =>
-      this.getConstantNode(toScopedIdentifier([...this.scopePath, this.name, enumerator])),
-    );
-  }
-
-  public toIdlMessageDefinition(): IdlMessageDefinition {
-    const definitions = this.enumeratorNodes().map((enumerator) =>
-      enumerator.toIDLMessageDefinitionField(),
-    );
-    return {
-      name: toScopedIdentifier([...this.scopePath, this.name]),
-      definitions,
-      // Going to use the module aggregated kind since that's what we store constants in
-      aggregatedKind: "module",
-    };
-  }
-}
-
 /** Takes a potentially scope-less name used in a given scope and attempts to find the Node in the map
  * that matches it's name in the local scope or any larger encapsulating scope.
  * For example:
@@ -193,6 +105,6 @@ function resolveScopedOrLocalNodeReference({
 
   return referencedNode;
 }
-function toScopedIdentifier(path: string[]): string {
+export function toScopedIdentifier(path: string[]): string {
   return path.join("::");
 }
