+++ packages/omgidl-parser/src/IdlNodes/index.ts
@@ -1,4 +1,8 @@
+export * from "./ConstantIdlNode";
+export * from "./EnumIdlNode";
 export * from "./IdlNode";
 export * from "./ModuleIdlNode";
 export * from "./ReferenceTypeIdlNode";
 export * from "./StructIdlNode";
+export * from "./StructMemberIdlNode";
+export * from "./TypedefIdlNode";
